requirements.txt
blinker==1.9.0
click==8.1.7
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
Werkzeug==3.1.3
redis==5.2.0
psycopg2-binary==2.9.10
flask-cors==5.0.0  # Add the correct version of flask-cors that you're using

